Move objects and classes to a proper file
