+++ app/schemas/com.eillia.ehya.model.data.database.AppDatabase/4.json
@@ -1,7 +1,7 @@
 {
   "formatVersion": 1,
   "database": {
-    "version": 3,
+    "version": 4,
     "identityHash": "b93b1d7fe436c895ca79e01cfb6b6784",
     "entities": [
       {
